Paste `NativeProxyCommon` into `NativeProxy` without thinking
